requirements.txt
# Core Data Science
pandas==2.0.3
numpy==1.24.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0
kmodes==0.12.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# MLFlow
mlflow==2.9.1

# Configuration
pyyaml==6.0

# Geographic
haversine==2.7.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
flake8==6.1.0
black==23.9.1
isort==5.12.0

# Environment
python-dotenv==1.0.0

# Streamlit (optional, for dashboard)
streamlit==1.28.0

# Documentation
sphinx==7.2.5
